estoy cansado jefe

+++ ContadSP/Models/DetalleProvision.cs
@@ -8,12 +8,12 @@ namespace ContadSP.Models
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int id { get; set; }
-        public int cantidad { get; set; }
+        public double cantidad { get; set; }
         public double subtotal_aprox { get; set; }
         public string? subtotal_letra { get; set; }
         public string? especificacion { get; set; }
         public bool presupuestado { get; set; } = false;
-        public int cantidad_aprobada { get; set; }
+        public double cantidad_aprobada { get; set; }
         public double subtotal_aprobado { get; set; }
         public string? subtotal_aprobado_letra { get; set; }
         public List<PresupuestoPedido> PresupuestoPedido { get; set; }
